feat: add post list request & search

+++ src/main/java/com/my/blog/dto/PostsDto.java
@@ -4,5 +4,5 @@ import com.my.blog.model.PostModel;
 
 import java.util.List;
 
-public record PostsDto(List<PostModel> posts, boolean hasPrev, boolean hasNext, int lastPage) {
+public record PostsDto(List<PostModel> posts, boolean hasPrev, boolean hasNext, long lastPage) {
 }
