+++ src/main/java/com/my/blog/model/PostModel.java
@@ -2,6 +2,6 @@ package com.my.blog.model;
 
 import java.util.List;
 
-public record PostModel(long id, String title, String text, List<String> tags, int likesCount, int commentsCount) {
+public record PostModel(long id, String title, String text, List<String> tags, long likesCount, long commentsCount) {
 
 }
